Refactor: Remove unused modules and types related to Trojan asteroids and universe generation
- Deleted `types.rs`, `builder.rs`, `dynamics.rs`, `mod.rs`, and `objects.rs` from the `trojans_asteroid` module.
- Removed `builder.rs`, `cosmic_time.rs`, `properties.rs`, and `mod.rs` from the `universe` module.
- Eliminated the `SystemType` enum and related structures from `stellar_systems/types.rs`.
- Cleaned up the main `stellar_objects.rs` file by consolidating and organizing core structures for better clarity and maintainability.
- Updated the RON serialization logic to reflect the removal of obsolete structures.

+++ src/physics/constants/common.rs
@@ -1,24 +1,24 @@
 /// Gravitationskonstante (m³ kg⁻¹ s⁻²)
-pub const G: f64 = 6.67430e-11;
+pub const G: f32 = 6.67430e-11;
 
 /// Lichtgeschwindigkeit in m/s
-pub const SPEED_OF_LIGHT: f64 = 2.99792458e8;
+pub const SPEED_OF_LIGHT: f32 = 2.99792458e8;
 
 /// Stefan-Boltzmann Konstante (W m⁻² K⁻⁴)
-pub const STEFAN_BOLTZMANN: f64 = 5.670374419e-8;
+pub const STEFAN_BOLTZMANN: f32 = 5.670374419e-8;
 
 /// Planck-Konstante (J⋅s)
-pub const PLANCK_CONSTANT: f64 = 6.62607015e-34;
+pub const PLANCK_CONSTANT: f32 = 6.62607015e-34;
 
 /// Boltzmann-Konstante (J/K)
-pub const BOLTZMANN_CONSTANT: f64 = 1.380649e-23;
+pub const BOLTZMANN_CONSTANT: f32 = 1.380649e-23;
 
 /// Mathematische Konstanten
-pub const PI: f64 = std::f64::consts::PI;
-pub const TAU: f64 = 2.0 * PI;
+pub const PI: f32 = std::f32::consts::PI;
+pub const TAU: f32 = 2.0 * PI;
 
 /// Umwandlung von Grad zu Radiant
-pub const DEG_TO_RAD: f64 = PI / 180.0;
+pub const DEG_TO_RAD: f32 = PI / 180.0;
 
 /// Umwandlung von Radiant zu Grad
-pub const RAD_TO_DEG: f64 = 180.0 / PI;
+pub const RAD_TO_DEG: f32 = 180.0 / PI;
