+++ src/physics/constants/stellar.rs
@@ -1,6 +1,6 @@
 /// Minimum Massenverhältnis für stabile L4/L5 Lagrange-Punkte
 /// Aus dem Artikel: Stern muss mindestens 24.96 mal schwerer sein als Planet
-pub const MIN_LAGRANGE_MASS_RATIO: f64 = 24.96;
+pub const MIN_LAGRANGE_MASS_RATIO: f32 = 24.96;
 
 /// Standard Epoch für astronomische Berechnungen (J2000.0)
 /// 12:00:00 TT on 1 January 2000 = JD 2451545.0
